refactor(android): init input package
