+++ liushu-android/app/src/main/java/com/elliot00/liushu/service/ImeService.kt
@@ -27,13 +27,13 @@ import androidx.savedstate.SavedStateRegistry
 import androidx.savedstate.SavedStateRegistryController
 import androidx.savedstate.SavedStateRegistryOwner
 import androidx.savedstate.setViewTreeSavedStateRegistryOwner
-import com.elliot00.liushu.KeyboardView
+import com.elliot00.liushu.input.InputView
 import com.elliot00.liushu.uniffi.Engine
 
 class ImeService : LifecycleInputMethodService(), ViewModelStoreOwner, SavedStateRegistryOwner {
     val engine = Engine()
     override fun onCreateInputView(): View {
-        val view = KeyboardView(this)
+        val view = InputView(this)
 
         window?.window?.decorView?.let { decorView ->
             decorView.setViewTreeLifecycleOwner(this)
